Created OData reading tests executor
