+++ tests/dummy/app/models/ember-flexberry-dummy-comment-vote.js
@@ -8,7 +8,7 @@ var Model = Projection.Model.extend({
     inverse: 'userVotes',
     async: false
   }),
-  
+
   //voteType: DS.attr('ember-flexberry-dummy-vote-type'),
 
   // This property is for flexberry-lookup component. No inverse relationship here.
