+++ tests/unit/query/odata-reading-restrictions-on-fields-test.js
@@ -1,88 +1,65 @@
 import Ember from 'ember';
-import { module, test } from 'qunit';
-
 import QueryBuilder from 'ember-flexberry-data/query/builder';
-import ODataAdapter from 'ember-flexberry-data/adapters/odata';
 import { DetailPredicate, StringPredicate } from 'ember-flexberry-data/query/predicate';
+import executeTest from './execute-odata-reading-test';
 
-import startApp from '../../helpers/start-app';
-import config from '../../../../dummy/config/environment';
+executeTest('reading | restrictions | on fields', (store, assert) => {
+  assert.expect(6);
+  let done = assert.async();
 
-if (config.APP.testODataService) {
-  const randKey = Math.floor(Math.random() * 9999);
-  const baseUrl = 'http://rtc-web:8081/odatatmp/ember' + randKey;
-  const app = startApp();
-  const store = app.__container__.lookup('service:store');
+  Ember.run(() => {
+    initTestData(store)
 
-  const adapter = ODataAdapter.create();
-  Ember.set(adapter, 'host', baseUrl);
+      // Reading by master field.
+      .then((records) => {
+        let authorId = records[0];
+        let builder = new QueryBuilder(store, 'ember-flexberry-dummy-comment')
+          .where('author.id', '==', authorId);
 
-  store.reopen({
-    adapterFor() {
-      return adapter;
-    }
+        return store.query('ember-flexberry-dummy-comment', builder.build())
+          .then((data) => {
+            assert.equal(data.get('length'), 2, 'Reading by master field | Length');
+            assert.ok(data.every(item => item.get('author.name') === 'Vasya'),
+              'Reading by master field | Data');
+            return records;
+          });
+      })
+
+      // Reading with master restrictions.
+      .then((records) => {
+        let commentId = records[1];
+
+        let sp1 = new StringPredicate('author.name').contains('Kolya');
+        let sp2 = new StringPredicate('text').contains('Comment 3');
+        let builder = new QueryBuilder(store, 'ember-flexberry-dummy-comment')
+          .where(sp1.and(sp2));
+
+        return store.query('ember-flexberry-dummy-comment', builder.build())
+          .then((data) => {
+            assert.equal(data.get('length'), 1, 'Restrictions on master fields | Length');
+            assert.ok(data.get('firstObject').get('id') === commentId,
+              'Restrictions on master fields | Data');
+          });
+      })
+
+      // Reading with detail restrictions.
+      .then(() => {
+        let dp = new DetailPredicate('userVotes')
+          .any(new StringPredicate('applicationUser.name').contains('Oleg'));
+        let builder = new QueryBuilder(store, 'ember-flexberry-dummy-comment')
+          .where(dp);
+
+        return store.query('ember-flexberry-dummy-comment', builder.build())
+          .then((data) => {
+            assert.equal(data.get('length'), 2, 'Restrictions on details fields | Length');
+            assert.ok(data.every(item => item.get('author.name') === 'Vasya'),
+              'Restrictions on details fields | Data');
+          });
+      })
+      .catch(e => console.log(e, e.message))
+      .finally(done);
   });
-
-  module('OData');
-
-  test('reading | restrictions | on fields', (assert) => {
-    assert.expect(6);
-    let done = assert.async();
-
-    Ember.run(() => {
-      initTestData(store)
-
-        // Reading by master field.
-        .then((records) => {
-          let authorId = records[0];
-          let builder = new QueryBuilder(store, 'ember-flexberry-dummy-comment')
-            .where('author.id', '==', authorId);
-
-          return store.query('ember-flexberry-dummy-comment', builder.build())
-            .then((data) => {
-              assert.equal(data.get('length'), 2, 'Reading by master field | Length');
-              assert.ok(data.every(item => item.get('author.name') === 'Vasya'),
-                'Reading by master field | Data');
-              return records;
-            });
-        })
-
-        // Reading with master restrictions.
-        .then((records) => {
-          let commentId = records[1];
-
-          let sp1 = new StringPredicate('author.name').contains('Kolya');
-          let sp2 = new StringPredicate('text').contains('Comment 3');
-          let builder = new QueryBuilder(store, 'ember-flexberry-dummy-comment')
-            .where(sp1.and(sp2));
-
-          return store.query('ember-flexberry-dummy-comment', builder.build())
-            .then((data) => {
-              assert.equal(data.get('length'), 1, 'Restrictions on master fields | Length');
-              assert.ok(data.get('firstObject').get('id') === commentId,
-                'Restrictions on master fields | Data');
-            });
-        })
-
-        // Reading with detail restrictions.
-        .then(() => {
-          let dp = new DetailPredicate('userVotes')
-            .any(new StringPredicate('applicationUser.name').contains('Oleg'));
-          let builder = new QueryBuilder(store, 'ember-flexberry-dummy-comment')
-            .where(dp);
-
-          return store.query('ember-flexberry-dummy-comment', builder.build())
-            .then((data) => {
-              assert.equal(data.get('length'), 2, 'Restrictions on details fields | Length');
-              assert.ok(data.every(item => item.get('author.name') === 'Vasya'),
-                'Restrictions on details fields | Data');
-            });
-        })
-        .catch(e => console.log(e, e.message))
-        .finally(done);
-    });
-  });
-}
+});
 
 function initTestData(store) {
   // Creates data for tests and returns neccessary data for them (2 ids).
@@ -115,7 +92,7 @@ function initTestData(store) {
   ])
 
     // Сreating suggestion.
-    .then((sugAttrsValues) => 
+    .then((sugAttrsValues) =>
       store.createRecord('ember-flexberry-dummy-suggestion', {
         type: sugAttrsValues.find(item => item.get('name') === 'Type 1'),
         author: sugAttrsValues.find(item => item.get('name') === 'Vasya'),
@@ -123,7 +100,7 @@ function initTestData(store) {
       }).save()
 
         // Creating comments.
-        .then((sug) => 
+        .then((sug) =>
           Ember.RSVP.Promise.all([
             store.createRecord('ember-flexberry-dummy-comment', {
               author: sugAttrsValues.find(item => item.get('name') === 'Vasya'),
@@ -151,7 +128,7 @@ function initTestData(store) {
           ])
 
             // Creating votes.
-            .then((comments) => 
+            .then((comments) =>
               Ember.RSVP.Promise.all([
                 store.createRecord('ember-flexberry-dummy-comment-vote', {
                   applicationUser: sugAttrsValues.find(item => item.get('name') === 'Oleg'),
