+++ tests/unit/query/odata-reading-restrictions-with-odata-func-test.js
@@ -1,60 +1,37 @@
 import Ember from 'ember';
-import { module, test } from 'qunit';
-
 import QueryBuilder from 'ember-flexberry-data/query/builder';
-import ODataAdapter from 'ember-flexberry-data/adapters/odata';
+import executeTest from './execute-odata-reading-test';
 
-import startApp from '../../helpers/start-app';
-import config from '../../../../dummy/config/environment';
+executeTest('reading | restrictions | odata functions', (store, assert) => {
+  assert.expect(4);
+  let done = assert.async();
 
-if (config.APP.testODataService) {
-  const randKey = Math.floor(Math.random() * 9999);
-  const baseUrl = 'http://rtc-web:8081/odatatmp/ember' + randKey;
-  const app = startApp();
-  const store = app.__container__.lookup('service:store');
+  Ember.run(() => {
+    initTestData(store)
 
-  const adapter = ODataAdapter.create();
-  Ember.set(adapter, 'host', baseUrl);
+      // User has a birthday tommorow.
+      .then(() => {
+        let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+          .where('birthday', '>', 'now()');
+        return runTest(store, builder, (data) => {
+          assert.ok(data.get('firstObject.name') === 'User 1', '> now() | Data');
+          assert.equal(data.get('length'), 1, '> now() | Length');
+        });
+      })
 
-  store.reopen({
-    adapterFor() {
-      return adapter;
-    }
+      // User had a birthday yesterday.
+      .then(() => {
+        let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+          .where('birthday', '<', 'now()');
+        return runTest(store, builder, (data) => {
+          assert.ok(data.get('firstObject.name') === 'User 2', '< now() | Data');
+          assert.equal(data.get('length'), 1, '< now() | Length');
+        });
+      })
+      .catch(e => console.log(e, e.message))
+      .finally(done);
   });
-
-  module('OData');
-
-  test('reading | restrictions | odata functions', (assert) => {
-    assert.expect(4);
-    let done = assert.async();
-
-    Ember.run(() => {
-      initTestData(store)
-
-        // User has a birthday tommorow.
-        .then(() => {
-          let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
-            .where('birthday', '>', 'now()');
-          return runTest(store, builder, (data) => {
-            assert.ok(data.get('firstObject.name') === 'User 1', '> now() | Data');
-            assert.equal(data.get('length'), 1, '> now() | Length');
-          });
-        })
-
-        // User had a birthday yesterday.
-        .then(() => {
-          let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
-            .where('birthday', '<', 'now()');
-          return runTest(store, builder, (data) => {
-            assert.ok(data.get('firstObject.name') === 'User 2', '< now() | Data');
-            assert.equal(data.get('length'), 1, '< now() | Length');
-          });
-        })
-        .catch(e => console.log(e, e.message))
-        .finally(done);
-    });
-  });
-}
+});
 
 function initTestData(store) {
   let tomorrowDate = new Date();
