+++ tests/unit/query/odata-reading-string-predicates-test.js
@@ -1,74 +1,51 @@
 import Ember from 'ember';
-import { module, test } from 'qunit';
-
 import QueryBuilder from 'ember-flexberry-data/query/builder';
-import ODataAdapter from 'ember-flexberry-data/adapters/odata';
 import { StringPredicate } from 'ember-flexberry-data/query/predicate';
+import executeTest from './execute-odata-reading-test';
 
-import startApp from '../../helpers/start-app';
-import config from '../../../../dummy/config/environment';
+executeTest('reading | predicates | string predicates', (store, assert) => {
+  assert.expect(4);
+  let done = assert.async();
 
-if (config.APP.testODataService) {
-  const randKey = Math.floor(Math.random() * 9999);
-  const baseUrl = 'http://rtc-web:8081/odatatmp/ember' + randKey;
-  const app = startApp();
-  const store = app.__container__.lookup('service:store');
+  Ember.run(() => {
+    initTestData(store)
 
-  const adapter = ODataAdapter.create();
-  Ember.set(adapter, 'host', baseUrl);
+    // Contains.
+    .then(() => {
+      let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+      .where(new StringPredicate('name').contains('as'));
 
-  store.reopen({
-    adapterFor() {
-      return adapter;
-    }
+      return store.query('ember-flexberry-dummy-application-user', builder.build())
+        .then((data) => {
+          assert.ok(data.every(item => item.get('name') === 'Vasya'),
+            'Contains with correct data | Data');
+          assert.equal(data.get('length'), 2, 'Contains with correct data | Length');
+        });
+    })
+
+    .then(() => {
+      let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+      .where(new StringPredicate('name').contains(null));
+
+      return store.query('ember-flexberry-dummy-application-user', builder.build())
+        .then((data) => {
+          assert.equal(data.get('length'), 0, 'Contains without data');
+        });
+    })
+
+    .then(() => {
+      let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+      .where(new StringPredicate('name').contains('Ge'));
+
+      return store.query('ember-flexberry-dummy-application-user', builder.build())
+        .then((data) => {
+          assert.equal(data.get('length'), 0, `Contains mustn't return any records`);
+        });
+    })
+    .catch(e => console.log(e, e.message))
+    .finally(done);
   });
-
-  module('OData');
-
-  test('reading | predicates | string predicates', (assert) => {
-    assert.expect(4);
-    let done = assert.async();
-
-    Ember.run(() => {
-      initTestData(store)
-
-      // Contains.
-      .then(() => {
-        let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
-        .where(new StringPredicate('name').contains('as'));
-
-        return store.query('ember-flexberry-dummy-application-user', builder.build())
-          .then((data) => {
-            assert.ok(data.every(item => item.get('name') === 'Vasya'),
-              'Contains with correct data | Data');
-            assert.equal(data.get('length'), 2, 'Contains with correct data | Length');
-          });
-      })
-
-      .then(() => {
-        let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
-        .where(new StringPredicate('name').contains(null));
-
-        return store.query('ember-flexberry-dummy-application-user', builder.build())
-          .then((data) => {
-            assert.equal(data.get('length'), 0, 'Contains without data');
-          });
-      })
-
-      .then(() => {
-        let builder = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
-        .where(new StringPredicate('name').contains('Ge'));
-
-        return store.query('ember-flexberry-dummy-application-user', builder.build())
-          .then((data) => {
-            assert.equal(data.get('length'), 0, `Contains mustn't return any records`);
-          });
-      })
-      .catch(e => console.log(e, e.message))
-      .finally(done);
-    });
-  });
-}
+});
 
 function initTestData(store) {
   return Ember.RSVP.Promise.all([
